requirements.txt
Django~=3.2
django-admin-sortable2~=1.0.3
Pillow~=9.0.1
django-tinymce~=3.4.0
